--- mentor-match-infra/requirements.txt
@@ -1,1 +1,0 @@
-aws-cdk-lib==2.79.1
