+++ mentor_match_infra/requirements.txt
@@ -1,0 +1,1 @@
+aws-cdk-lib==2.79.1
